Extract ConsoleIO effect
